function/requirements.txt
azure-functions
psycopg2-binary==2.9.6
sendgrid==6.10.0